ARM3 64bit portability fixes.
svn path=/branches/ros-amd64-bringup/; revision=43428

+++ reactos/ntoskrnl/mm/ARM3/contmem.c
@@ -428,7 +428,7 @@ MmFreeContiguousMemory(IN PVOID BaseAddress)
 VOID
 NTAPI
 MmFreeContiguousMemorySpecifyCache(IN PVOID BaseAddress,
-                                   IN ULONG NumberOfBytes,
+                                   IN SIZE_T NumberOfBytes,
                                    IN MEMORY_CACHING_TYPE CacheType)
 {
     //
